requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
discord==2.3.2
discord.py==2.3.2
exceptiongroup==1.2.1
fake-useragent==1.5.1
filelock==3.13.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
jmespath==1.0.1
keepa==1.3.9
lxml==5.2.1
lxml_html_clean==0.1.1
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parse==1.20.1
parsel==1.9.1
playwright==1.43.0
playwright-stealth==1.0.6
pyaes==1.6.1
pyasn1==0.6.0
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-html==0.10.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
Telethon==1.35.0
tldextract==5.1.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
unshortenit==0.4.0
urlexpander==0.0.37
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
word2number==1.1
yarl==1.9.4
zipp==3.18.1
